make type of SymbolSubscribable global
While the value of SymbolSubscribable was global, the type was not,
and so TypeScript did not recognize the different versions of
`SymbolSubscribable` from different package version as being the same
type. As a result, the `Subscribable` interface:

```ts
interface Subscribable<T,TResult> {
  [SymbolSubscribable]); Operation<Subscription<T,TResult>;
}
```

while sharing a name, had a different structure and so TypeScript
effectively recognized them as a different interface from different
package versions. So a subscribable defined with
"@effection/subscription" 0.8.0 and a function that consumed
subscribables using the definion in 0.8.1 did not type-match.

This allocates a totally fake global variable whose unique type can be
used as the type of `SymbolSubscribable`. That way, TypeScript will
always use that global type no matter which version of the package it
is using and so it will recognie interfaces that reference that symbol
as being the same.