Make `createSubscription` chainable
Subscriptions created via `createSubscription` are chainable on both sides of the yield statement. This enables patterns like:

```
yield on(window, "message").filter((m) => ...).forEach(...)
```

There is a subtle downside and backwards incompatibility with this, which is that due to `Operation` being invariant over its type parameter. Since we return a `Operation<ChainableSubscription<T>>` but we sometimes expect a `Operation<Subscription<T>>` and while `ChainableSubscription <: Subscription` really *should* imply that `Operation<ChainableSubscription> <: Operation<Subscription>`, this is currently not (always) the case, so some subtle type errors might occur.