Introduce cross-platform process API
We have had a lot of pain both in representing our processes in not
only a cross platform way, but also trying to use a "one size fits
all" approach for resource management. Sometimes you want to
implicitly attach an error handler to a process and fail if it ever is
triggered. Other times, you want to handle failure explicitly so you
can do things like restart, etc.. Our existing process abstraction was
too coupled to low-level node constructs that are already frustrating
ad fiddly to begin with and so it seriously constrained what we

This takes a "less is more" approach of defining a very thin,
effection-friedly API for process management that takes care of the
following cases:

1. running a process to completion, and failing with an exception if
that process does not complete.
2. running a process to completion and handling the output manually
3. running a process forever as you would a web server, and failing
with an error if it exits for _any reason whatsoever_

For cases (1) and (2) this introduces the `exec()` operation which
produces an instance of the `Process` api for finite lengt process
that are expected to quit at some point. Key to the refactor is that
errors in the process do not necessarily cause an error to be
raised. Instead, they are collected into an `errors[]` array that is
part of the process `ExitStatus` structure.

For Windows support, we take the strategy of allowing a completely
parallel implementation ofy the `Process` API that is
not (necessarily) shared between implementations. So when you call
`exec()` windows implementation has complete freedom on how to manage
the child process internally.

With `exec()` in place, we then implement a `daemon()` operation to
handle case (3).

```ts
yield daemon("node http-server");
```

will spin up the command and will raise an error inside the operation
context if the http server fails for whatever reason
whatsoever. Naturally, processes created with both `exec()` and
`daemon()` will be shut down if they ever pass out of scope.