✨add pid to Process interface
In order to make our test cases cross platform, we need to be able to,
from the outside in, send a signal to the process (or kill it using a
custom windows module) In both of those cases, we need a process id.

This adds the `pid` to the Process api and into the posix
implementation. This did mean that the process had to be created
outside of the resource body, but that should be ok.