Add support for win32
We want windows developers to feel right at home using Effection and
BigTest, and so that means that the same liberating guarantees about
process shutdown need to apply on the win32 platform as well as just
POSIX platforms.

This provieds a parallel implementation of the `exec()` api for
Windows that has windows specific hooks. It follows closely the posix
implementation, with a few key differences.

1. cross-spawn is used to create the ChildProcess since it handles
automatic resolution of `.cmd`, `.bat`, `.exe` files etc.
2. On closing the process we make sure to close the `stdin` handle,
and also send a ctrl-c event to it using the `ctrlc-windows` package.

It's worth noting that we may need to take extra measures in the
future to make sure that the child process is spawned with its own
console since the ctrlc-windows will not only send a ctrlc event to
the target process, but also will send it to every process that shared
that process's console.