Remove generators as a valid operation type
Currently we allow both generator functions and instantiated generators
as operation types. This change removes support for instantiated
generators entirely.

The reason to do this is that being able to pass a generator is
confusing and inconsistent. We've actually seen many strange bugs due
to this in BigTest, where simply changing to a generator function fixed
the issue. The problem is that generators are not *sharable*. A
generator being polled by two tasks at the same time will cause
indeterminate results. A generator function on the other hand can safely
be invoked by both tasks.

This also creates consistency, where the first argument to a generator
function is *always* the task, and we are not tempted to pass arguments
to the generator function itself, like with the `sleep` function which
has been updated in this PR.