Make generators reentrant
A task may be currently running a generator, that is it has entered the
generator and is currently inside the generator itself in between two
yield points.

Since resume/reject and halting is now fundamentally synchronous, it
could happen that the code in the generator itself causes the task to
halt, either by explicitly halting itself, or by causing a sibling or
parent task to fail and thereby being halted by the parent.

However, this poses a problem, since the iterator is already running, we
cannot call `return()` on it. This will cause a `TypeError: Generator is
already running`.

Effectively we cannot enter a generator if we have already entered the
generator.

This stores a list of continuations, so that we can effectively delay
the halt until we are done running the generator itself.