Add `stringBuffer` method to `Stream`
This method retains a string which concatenates all previously emitted
values to each other. This should normally be used on streams of strings
or node buffers, however we cannot constrain this method to only work on
these types (unless we make it a free function instead). This type of
stream is useful for use with node io streams, to aggregate their
results in memory. We will be able to use this with the io streams in
`@effection/node` to make working with them far easier.