Make mocha a peer dependency for `@effection/mocha`
@effection/mocha had a hard dependency on `mocha` which meant that it
had its own local copy of mocha in its `node_modules/` directory. When
there was only ever one copy of mocha in the NODE_PATH, that was fine
because the runtime object required was the same every time, and the
global `beforEach` hooks could be used. However, if there were
different versions or even different copies of the same version, the
mocha module required by `@effection/mocha` was "inert" meaning that
because it wasn't the one being run, the global hooks were not even
present.

This sets mocha as a peer dependency so that `@effection/mocha` will
_never_ have its own local copy of `mocha`. There is still the
possibility of more than one mocha in the path and of
`@effection/mocha` getting a different copy than the one that is
currently running, but we can cross that bridge when we come to it.