Simplify Atom implementation and remove dependencies
Our Atom has always been based on lenses and functional programming.
However, with the advent of Streams in effection, I have been thinking
about a simpler implementation of Atom. I realized that we could make an
implementation which stacks slices on top of each other, rather than
accessing the state directly. Effectively our slice becomes sort of like
a lens itself, except with the addition of a stream.

Since there were some chicken-and-the-egg problems with making slices, I
introduced `Container` as a simplified slice, this also makes it easy to
attach all of the additional method for slice which are not needed for
the traversal.

In my opinion the resulting implementation is easier to understand, and
also means we do not need to depend on any external packages in
`@effection/atom`.

The resulting code does make rather judicious use of `any` to skip
around the inherent difficulty in making a type safe lens-like thing in
TypeScript. However, since the type safety of Slices is provided by the
`MakeSlice` type, there is no loss in external type safety.