Disable command parsing when exec({ shell: true })
> resolves #266

The current strategy of using `shellwords` to parse command and
arguments works for single commands. However, for complex commands
that do piping and IO redirection, it does not work. But it is
precisely for this case that the `shell: true` option exists.

In the event that `shell: true` is set, we do no parsing at all and
just pass the command wholesale down to let the shell do its own
parsing.