Store effection root in global variable
If there are multiple versions of the effection package, for example if
effection is also a transitive dependency, then we might end up with
multiple roots. By using a global variable we can share the same root
between all instances of effection. The name of the global variable
includes the major version of effection, so that we can make breaking
changes to its layout in major versions.