Add once() equivalent
The ability to synchronize on an observed condition is extremely
handy, and we use it quite often in `Atom`. It would be nice if we
could define this upstream in the stream interface instead of the
atom.

Originally, I set about adding a `once()` method for stream that was
just a composition of `filter(predicate).first()`. But there are cases
that you want to have the actual matched value, and not just
synchronize on it:

```ts
let match = yield stream.once(({ status }) => status === 'ready');
```

In those cases where you want the return value, this method should
behave like `expect()`, raising an exception if the stream is closed
before a value is received.

On other occasions however, you don't necessarily need the value, you
just want to synchronize, and in those cases it's ok for the stream to
close. In order to support both of these cases, we would need both a
`once()` and `onceExpect()` so I opted instead to allow the `first()`
and `expect()` methods to accept a predicate against which to
match. Thus, first really means "first matching predicate" and expect
means "expect a value matching predicate".