Lock dependency versions while in prerelease mode
Since minor and patch versions are equivalent in prerelease mode, and
the caret `^` operator does not work well in prerelease mode, we should
lock the versions of internal dependencies.

Changesets will keep the type of dependency, so when the version is
bumped it will keep it as a strict dependency.

Since bumping a package will also release all of its downstream
packages, it should be easier to upgrade packages together.