Emit initial state as part of atom stream, deprecate once.
As part of simplifying the atom, it was decided to make the atom
implement the streaming API directly. However, one of the extremely
useful cases for atom is to synchronize on state. It is really nice
because it impervious to the same race conditions that event listeers
are. This is because you can match against the current state and all
future states and so it does not matter when you start your
subscription, because you are guaranteed that _if_ the "event" that
you're looking for arises, you will see it. Even better, there is zero
additional memory cost that would come with solution like buffering
because you're only matching against an already allocated data
structure.

It's desirable to not have to use a one-off API when the same can be
achieved my composing `filter()` and `expect()`. However, in order to
achieve this, but also retain the semantics of the `once()` method, we
have to be able to filter on the _current_ atom state as well.

This creates a new stream off of the channel of states and publishes
the current state first, before publishng.

The `once()` method, now no longer needed has been deprecated. It will
issue a single warning per atom, which is multiple in the test suite,
but only once per application.