Make Queue a first class citizen
The fundamental primitive which implements the subscription interface is
`Queue`. This realization means that we should make queue a first class
citizen.  This makes `createQueue` and the `Queue` interface, makes
`Queue` implement `Subscription` and makes it closable.

We keep `createSubscription` for parity with Stream, but it is simply
implemented on top of `Queue`.

For now we will keep the `OperationIterator` and `OperationIterable`
primitives, but I do not see them really pulling their weight, since
using rich objects is so much more convenient. In the future we might
want to get rid of these as well.