Run destructors in reverse order and in series
When spawning multiple children in a task, it is intuitive that the
child that was spawned will be terminated first (last in first out),
this makes sense for example if the second child has some dependency on
the first (e.g. uses its network services), and we expect the second
child to be cleaned up before the resource it depends on is cleaned up.

For the same reason, destructors must be run in *series* rather than in
parallel. Termination of the second child must complete *fully* before
even initiating termination of the first. This may sound very
inefficient at first glance, but it is not as bas as it seems, since a
vast majority of children in fact terminate synchronously, and async
termination is an edge case.

The implementation strategy has been to reuse the *trapper* mechanism
and add an additional trapper to the child which continues the
termination until there are no haltable children left. There is a
slightly tricky order-dependency here in that the child must have been
removed from the set before the added trapper triggers. This is safe for
now since: 1) trappers are executed in order and 2) the trapper which
removes the child from the set is added on `link` which must always be
called before halt, so there is no possibility of these executing out of
order. Nevertheless, this implicit dependency is a little worrisome.