Allow for pending it blocks
Mocha allows for `it()` blocks that do not yet have a body. This is so
that you can write out the flow of a testcase before you actually get
around to making it work. To support this, the second argument to
`mocha.it()` is actually optional.

```js
describe('a thing', () => {
  it('will do this');
  it('will do that');
})
```

This allows the same construction in `@effection/mocha`