✨Add a race() combinator
Called for in https://github.com/thefrontside/effection/issues/292

This adds a `race()` combinator that runs an array of operations in
parallel where the first one to complete becomes the result of the
race operation.

If any of the operations fail, then the race operation fails.

It mirrors the functionality of [Promise.race][1]

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/race