Remove delay before starting iterator
In v2 we started off with a lot of asynchronicity, but the more we did
this, the more we realized that synchronous operation is a core part of
the workings of effection. It prevents race conditions in criticial
portions of the code and allows us to interoperate with callbacks and
evented code seamlessly.

This is the last remaining piece of asynchronous execution, where when
we step into an iterator controller, we delay the start of execution
until the next tick. There is no technical reason to do this, aside from
the astheatic of delaying the tick.

Removing this delay puts us on a path towards getting rid of resolution
functions entirely, since we now can do both synchronous setup as well
as synchronous teardown with generators. This does require a bit of
caution when implementing, but this should be mostly restricted to
primitives, such as `once`.

Interestingly, the current behaviour has actually caused a potential
race condition in `once`, since it is now implemented on top of
`onceEmit` with a generator function. This PR fixes this issue without
having to change the code in `@effection/events`.