Replace resolution functions with futures
Futures represent a synchonously resolving future value. This is exactly
why we added resolution functions in the first place: So we can have a
leaf-node which resolves synchronously. While futures are slightly more
cumbersome to use than resolution functions, this makes everything more
consistent and reduces the number of involved concepts.

This allows us to treat tasks as futures when it comes to using them
as an operation, rather than treating them as promises like we did
before. The advantage of this is that if the task resolves
synchronously, then using it as an operation becomes synchronous as
well. This becomes important when `yield`-ing to a task.