Publish npm packages from `v0` branch using the `r0` tag.
It is very unlikely that we would publish an Effection npm package
from the `v0` branch, but as configured, this will make it the
`latest` version on npm which is almost certainly not what we want.

If, on the offhand chance that, that we would need to publish a patch
from the `v0` branch, it would be published under the `r0`
tags (standing for "release 0"). I would have like to have made this
`v0`, but tags cannot be valid semver values, and TIL semver specs may
optionally start with `v`, so that makes `v0` illegal.