Publish npm packages from `v1` branch using the `r1` tag.
It is very unlikely that we would publish an Effection npm package
from the `v1` branch, but as configured, this will make it the
`latest` version on npm which is almost certainly not what we want.

If, on the offhand chance that, that we would need to publish a patch
from the `v1` branch, it would be published under the `r1`
tags (standing for "release 1"). I would have like to have made this
`v1`, but tags cannot be valid semver values, and TIL semver specs may
optionally start with `v`, so that makes `v1` illegal.

Also, this disables pr previews since this branch is in maintenance mode.