Start beta at the beginning
Changesets was just swapping out the tag, but keeping the version
point numbers in place which means that the beta tag was not really
starting from anywhere meaningful.

This levels out all packages to start at 2.0.0-beta.0 by removing the
existing changsets. These are already in the changelogs, and so it's
ok if we drop them.