Draft for immediate publication
Now that `v2` is the default branch, and the latest tag. It's really
important to have _something_ that is representative of the current
state.

This reorganizes the structure to introduce the WHY question first,
and proceed from their. The tutorial and examples were compressed so
that they could build off of one another.

We can expand further, but didn't want this to sit over the weekend.