Simplify README
Now that we have a [website](https://frontside.com/effection), we
don't actually need to have major content inside the README since it
would be duplicating our efforts.

Instead, this removes most of the "howto" guide-style aspect of the
README, and instead just provides a very short introduction to
Effection, and then links to the documentation site.

A brief introduction to contribution is included, but it is expected
that it will be fleshed out more later