Allow anything labellable.
As part of making everything labelled, it became obvious that there
are certain things which are _not_ opeartions that need to be labelled
in order to make the operations that they give rise to be
labelled. For example, a `Stream` is not an operation, but
`stream.forEach` is. Therefore, we want to make sure we can name the
stream, and then label that stream's operations. E.g.

```
on('message').forEach()
```

This relaxes the constraint that the only thing that can be used with
`withLabels` is an Operation, and instead says, if it is labelable,
then you can label it.