Label the suspend operation
Currently in the inspector, whenever you are suspended indefinitely,
it just shows us as 'task'. This is not particularly helpful since
it's unclear what exactly is going on. Instead, we'd like a clear
indicator that the operation is a suspend operation and will never,
ever return by itself.

We handle this in the create task and check if the operation is
undefined, if so, we assign it the `suspend` label.