Inherit labels from delegate operations.
A function operation is a fundamentally delegating construct. It is a
function that takes a scope, and returns an operation that actually
does the work. Currently however, when this happens the labels of the
actual operation are lost if that operation is the destination of a
function operation.

In other words, if I have an operation:

```ts
run(() => ({
  name: 'console.log',
  labels: {
    platforms: ['node', 'browsers']
  },
  perform() {
    console.log("I log therefore I am");
  }
}));
```

This operation will _not_ actually have any labels associated with it
and will just show up as `task` in the inspector tree. Instead, we'd
like to have those labels of the actual operation available for
inspection.

We can fix this by having function controllers "lift" the labels of
their source operation into their own task. This is feasible because
when a function controller starts, the first thing it does is
recursively resolve the destination controller into a controller that
has a concrete operation.

This changes the `Controller` API to expose its operation, so that no
matter what kind of controller is delegated to by the function
controller, it can import the labels of that operation.