Add labels to streams, queues, and events
streams, and queues are core constructs that are completely anonymous
right now. This means that it's totally impossible to find out what's
going on when using them.

This let's you pass an optional `name` parameter to the
`createQueue()` and `createStream()` methods which will then be
prepended to each operation for those things. So if you create a
stream called `messages`, then the forEach operation on that stream
will be called `messages.forEach`.

Using this low level mechanism, we create names for the streams
created by `on()` and `onEmit()`.

In order to make the visualization clean, `on()`, and `once()` no
longer delegate to `onEmit()` and `onceEmit()`. While it makes the
implementation of the core library a little less clear, it makes the
tree a lot more compact and easier to understand. I think this is
worth the tradeoff since these are going to be very heavily used
operations.