update yarn.rock to reflect dependencies.
yarn.lock is not congruent with the collective versions in
packages.json

This just checks in a compatible version. However, it may be worth
looking into see if this is a one-off or whether the pull request
generated by covector is not updating the lockfile to point to the new
version debpendencies.