Add spawn operation directly to Task.
> resolves #360

This adds a `spawn()` operation to the `Task` api which,
unsuprisingly, spawns a child in the context of a particular
task. Note that this is an _operation_, and not a synchronous
function. Calling the `spawn()` method will not actually do anything
until the operation is evaluated.

Now, it is very easy and safe to spawn operations within the context
of a specific task without having to use the `within()` syntax (which
we might deprecate in the future).