Introduce task scope
This introduces the idea of task scope, as a lower level construct which
can be used instead of resources.

Currently, some base operations, such as `spawn` and `label` are based
on resources, because resources are the only escape hatch allowing one
to "expand the scope" as we like to say, or it can be seen as "inlining"
an operation.

However this has presented a problem for us: we want resources to
introduce a task which "groups" a resources tasks, which can be seen in
the inspector. But adding this additional layer is incompatible with
e.g. `spawn` being a resource. This solves the problem by introducing
the concept of task `scope`. Every task which is not a root task has a
scope. Normally this scope is the parent task, but it does not
necessarily have to be.

This scope is accessible to "inlining" operations such as `spawn`, and
also indirectly to resources, since the newly introduce resourceTask is
running inside the scope.

This also prepares the way for fixing an exisiting problem with
resources, namely higher order operations such as `all` which currently
do not work well with resources. An operation such as `all` can pass
down its own scope, instead of creating its children with *itself* as
the scope which would be the default. This will make something like the
following work as expected:

```
yield all([someResource])
```