📄Link README of effection package to main repo readme
On the main `npm` package of Effection, there is no README which means
if that is the first place that a person looks for more information,
they will be bounced from all our great documentation rather than
drawn in.

This just adds a symlink to the README.md from the root which is our
front page on GitHub. That way, we can have it be the front door for
both GitHub _and_ NPM