Use Object.create to create error object with trace
Unfortunately copying the properties to the error objects as we did
before does not work in all JS engines, notably Firefox, see #485

This option gives us a less clean prototype chain, but the properties
do seem to at least be correctly available on all JS engines.

Another option for implementing this would be using a proxy object,
but I don't think there is any benefit to doing so here.