Share event loop among controller, task and future
We currently have three different event loops for a single task, the one
in the future, the one in the task, and the one in the iterator
controller. This unifies all three event loops so we only have one loop.
Additionally there are some tweaks to which sections are protected by
the event loop. Hopefully this will solve some reentrancy issues, but it
is a little bit of a stab in the dark.