Catch errors thrown during resource init
Closes #535

This adjusts the propagation of errors for resources so that errors
thrown during `init` can be cought. Previously, an error thrown during
`init` would propagate to the resource task, causing the resource task
to fail. This would have the undesirable side-effect that even if the
`init` error was caught (for example to retry), then the outer task
would still fail, due to the failure in the resource task.

Instead, a failure in `init` now causes the resource task to become
*halted* rather than *errored*, sort of as though it were a sibling,
while still maintaining the existing hierarchy of the `init` task being
a child.

This means that just as before, a failure in `init` will cause the
resource to be shut down, but this failure will not propagate out to the
outer task. Since the failure will always be thrown from the *resource
constructor*, the failure will still propagate, but we have more
control, since we can catch the error from the resource constructor if
it is yielded to.