Remove unused `OperationIterable` interface
In olden times, we used a special symbol to determine if something was
"subscribable" so that when yielded to, it could take a task, and bind
a subscription to it. However, this mechanism has been superceded by
resources and as a result was not in use anywhere in the codebase. It
contained some really disgusting TypeScript hackery to make it work,
and so we should get it out of here before releasing.