Add typedocs for @effection/fetch
I also optimized the module exports while doing this to account for
two things:

1. The `Fetch` interface was not being exported from `fetch.ts` which
makes it difficult to properly type a reference to `fetch()` using
typescript. It also means that it wasn't showing up in the typedoc.
2. `AbortController` was being exported for no good reason that I can
determine. Any abort signal that you pass down into fetch is
ignored. As far as I can tell the only effect of exporting
`AbortController` is that it showed up in the typedoc.