Add `Sink` as a unifying interface for `send`
Adds a new interface `Sink` which specifies a `send` method as a
unifying interface for all things which can be sent messages to. It
functions as the opposite end of a subscription, basically a destination
for a subscription.

This makes `send` an operation in those cases where it is not, for
example for `Channel` and `Queue`. It also turns the `publish` callback
on `createStream` into an operation, as well as `set`, `update` and
`remove` on `Atom`. All of these are breaking changes.