Work around borked NPM release
While releasing all of the 2.0 packages, NPM had a techincal issue
that cause 7 of our 19 to be in an inconsistent state where they were
received, but not viewable. After contacting NPM, they said that the
old version, in this case 2.0.0 was not usable anymore.

This introduces a patch to @effection/core that does not accompany any
changes, but is just there to cause every other version to increase