Standardize build tools for all packages
While trying to build a @jest integration, I was getting unrelated
yarn errors because of an old version of parcel being incompatible
with a certain yarn version. We need to upgrade parcel in inspect-ui,
but first we need to make sure that we're using the name node and yarn
everywhere.

It turns out that there were several packages that were using their
own volta configs. This points all of them at the same config from the
base of the repo