Add @effection/jest to covector
In order to have the version numbers advance automatically with
changesets like the other packages.

The initial release of a package needs to be manual, but after that we
can add them to covector. This also adjusts the version of effection
that it depends on to be the most recent.