Add testing guide for Jest and Mocha
Our unit testing integration is one of the most powerful aspects of
Effection, but right now there is not much documentation on the
concepts of how to make it work.

This adds a general testing strategy for how to do testing with
effection, followed by a mini-tutorial on using it with Jest and
Mocha since we a) have those already and b) they are so close together
that we can speak both audiences of users together.