👍Convert references to "scope" to use "task"
In Effection, the actual thing is a task that manages scope, so in
order to be unambiguous, let's refer to task when we're talking about
the actual thing which is created, and halted.

In a couple of cases, I've called the task associated with the suite
the "suite-scoped" task, and the task with the lifetime of the `it`
block the "test-scoped" task. In both cases, the thing we're talking
about is a task.