Create totally independent scopes for before{All/Each}()
It turns out that you don't actually want to have the `beforeEach`
scope to descend from the `beforeAll` because if we have a failure in
that scope, the parent needs to outlive it so that it is around when
the second test runs. As it is now, if you have two testcases:

+ before all
   -> it does one thing
   -> it does another thing

If the first test fails, it will kill the before all scope because it
is a child and so the second test case fails.

This makes the suite scope and the test scope *independent siblings*
so that a test can fail without crashing the suite.

This also adds some checks to make sure both before and after a test
runs that both scopes are in good order and if they are failing, the
error is raised.