Don't run each it.eventually trial in test scope
it.eventually() runs the test operation over and over again until it
either passes, or times out. However, we were running each trial
directly in the test's "each" scope, so if the trial failed, then it
crashed the test's scope which then caused the test to crash, even
though we were catching the error inside the eventually loop.

This just calls the trial as a normal operation whithin the scope of
the current task, that way it can either catch or propagate using the
normal rules of nested tasks.