✨Build inspect-ui with stable parcel
The ESM entry points don't exist when trying to include the
inspect ui into another UI (in this case Backstage), so obviously this
needs to be fixed. However, parcel is in the way by using the
"browser" field of package.json. When you import from inspect ui in a
React app compiled with webpack, you don't want to require from
`index.html`, you want to get at the actual ESM bundled code. The
first step then, is to move `app.html` away from "browser" field and
just build it to a `dist-app` directory that can contain the entire
app. Then, we can build the JavaScript code to `dist-esm` and actually
make imports from there work.

This upgrades parcel to the next version and uses an "app" target to
build it.